Refactor import paths for AppSidebar in client chatroom, home, and profile pages
