+++ the-pack/src/app/(client)/clienthome/page.tsx
@@ -3,7 +3,7 @@
 import React from 'react';
 import ClientCarousel from "@/components/client/home/ClientCarousel"
 import { SidebarProvider, SidebarTrigger } from "@/components/ui/sidebar"
-import { AppSidebar } from "@/components/app-sidebar"
+import { AppSidebar } from "@/components/client/sidebar/app-sidebar"
 import Scheduler from '@/components/coach/home/Scheduler';
 import { Separator } from "@/components/ui/separator"
